// src/lib/trigram/type.ts
export type TrigramSymbol = '☰' | '☵' | '☶' | '☳' | '☴' | '☲' | '☷' | '☱';

export type TrigramType = '乾' | '坎' | '艮' | '震' | '巽' | '離' | '坤' | '兌';

export type TrigramFullName =
  | '乾為天'
  | '坎為水'
  | '艮為山'
  | '震為雷'
  | '巽為風'
  | '離為火'
  | '坤為地'
  | '兌為澤'
  | '乾為天';

export type TrigramElement =
  | '天'
  | '水'
  | '山'
  | '雷'
  | '風'
  | '火'
  | '地'
  | '澤'
  | '天';

export type TrigramNumber = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8;

export type TrigramIndex = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7;